# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deliverance Church Lugazi (DCL) Church Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /branches:
    get:
      tags: [Branches]
      operationId: listBranches
      summary: List all church branches
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Branch'
    post:
      tags: [Branches]
      operationId: createBranch
      summary: Create a branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
  /branches/{id}:
    get:
      tags: [Branches]
      operationId: getBranch
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
    patch:
      tags: [Branches]
      operationId: updateBranch
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Branch'
    delete:
      tags: [Branches]
      operationId: deleteBranch
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
...
  /departments:
    get:
      tags: [Departments]
      operationId: listDepartments
      parameters:
        - name: branchId
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
    post:
      tags: [Departments]
      operationId: createDepartment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
  /departments/{id}:
    get:
      tags: [Departments]
      operationId: getDepartment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
    patch:
      tags: [Departments]
      operationId: updateDepartment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
    delete:
      tags: [Departments]
      operationId: deleteDepartment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
  /announcements:
    get:
      tags: [Announcements]
      operationId: listAnnouncements
      parameters:
        - name: pinned
          in: query
          required: false
          schema:
            type: boolean
        - name: publicOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Announcement'
    post:
      tags: [Announcements]
      operationId: createAnnouncement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncementInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
  /announcements/{id}:
    get:
      tags: [Announcements]
      operationId: getAnnouncement
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
    patch:
      tags: [Announcements]
      operationId: updateAnnouncement
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncementInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
    delete:
      tags: [Announcements]
      operationId: deleteAnnouncement
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
...
  /attendance/services:
    get:
      tags: [Attendance]
      operationId: listAttendanceServices
      parameters:
        - name: branchId
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceService'
    post:
      tags: [Attendance]
      operationId: createAttendanceService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceServiceInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceService'
  /attendance/services/{id}:
    patch:
      tags: [Attendance]
      operationId: updateAttendanceService
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceServiceInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceService'
    delete:
      tags: [Attendance]
      operationId: deleteAttendanceService
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Branch:
      type: object
      required: [id, name, isMain, location, pastorInChargeName, contactPhone, contactEmail, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        isMain:
          type: boolean
        location:
          type: string
        pastorInChargeName:
          type: [string, 'null']
        contactPhone:
          type: [string, 'null']
        contactEmail:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BranchInput:
      type: object
      required: [name, location]
      properties:
        name:
          type: string
        isMain:
          type: boolean
        location:
          type: string
        pastorInChargeName:
          type: [string, 'null']
        contactPhone:
          type: [string, 'null']
        contactEmail:
          type: [string, 'null']
...
    Department:
      type: object
      required: [id, name, description, branchId, leaderId, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: [string, 'null']
        branchId:
          type: [integer, 'null']
        leaderId:
          type: [integer, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: [string, 'null']
        branchId:
          type: [integer, 'null']
        leaderId:
          type: [integer, 'null']
    Announcement:
      type: object
      required: [id, title, body, mediaUrl, mediaType, isPublic, isPinned, branchId, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        mediaUrl:
          type: [string, 'null']
        mediaType:
          type: [string, 'null']
          enum: [image, video, audio, null]
        isPublic:
          type: boolean
        isPinned:
          type: boolean
        branchId:
          type: [integer, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AnnouncementInput:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
        body:
          type: string
        mediaUrl:
          type: [string, 'null']
        mediaType:
          type: [string, 'null']
          enum: [image, video, audio, null]
        isPublic:
          type: boolean
        isPinned:
          type: boolean
        branchId:
          type: [integer, 'null']
...
    AttendanceService:
      type: object
      required: [id, title, branchId, serviceDate, adultCount, youthCount, childrenCount, totalCount, notes, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        branchId:
          type: [integer, 'null']
        serviceDate:
          type: string
          format: date
        adultCount:
          type: integer
        youthCount:
          type: integer
        childrenCount:
          type: integer
        totalCount:
          type: integer
        notes:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AttendanceServiceInput:
      type: object
      required: [title, serviceDate]
      properties:
        title:
          type: string
        branchId:
          type: [integer, 'null']
        serviceDate:
          type: string
          format: date
        adultCount:
          type: integer
        youthCount:
          type: integer
        childrenCount:
          type: integer
        notes:
          type: [string, 'null']
...